+++ code/frameworks/angular/src/client/decorators.test.ts
@@ -2,9 +2,9 @@ import { Addon_StoryContext } from '@storybook/types';
 
 import { Component } from '@angular/core';
 import { moduleMetadata } from './decorators';
-import { AngularFramework } from './types';
+import { AngularRenderer } from './types';
 
-const defaultContext: Addon_StoryContext<AngularFramework> = {
+const defaultContext: Addon_StoryContext<AngularRenderer> = {
   componentId: 'unspecified',
   kind: 'unspecified',
   title: 'unspecified',