+++ code/frameworks/angular/src/client/render.ts
@@ -1,13 +1,13 @@
 import { Store_RenderContext, ArgsStoryFn } from '@storybook/types';
 
 import { renderNgApp } from './angular/helpers';
-import { AngularFramework } from './types';
+import { AngularRenderer } from './types';
 
 import { RendererFactory } from './angular-beta/RendererFactory';
 
 export const rendererFactory = new RendererFactory();
 
-export const render: ArgsStoryFn<AngularFramework> = (props) => ({ props });
+export const render: ArgsStoryFn<AngularRenderer> = (props) => ({ props });
 
 export async function renderToCanvas(
   {
@@ -16,7 +16,7 @@ export async function renderToCanvas(
     forceRemount,
     storyContext: { parameters, component },
     id,
-  }: Store_RenderContext<AngularFramework>,
+  }: Store_RenderContext<AngularRenderer>,
   element: HTMLElement
 ) {
   showMain();