+++ code/lib/preview-web/src/PreviewWeb.test.ts
@@ -29,8 +29,7 @@ import {
 } from '@storybook/core-events';
 import { logger } from '@storybook/client-logger';
 import { addons, mockChannel as createMockChannel } from '@storybook/addons';
-import type { Framework } from '@storybook/types';
-import type { ModuleImportFn, WebProjectAnnotations } from '@storybook/store';
+import type { Renderer, Store_ModuleImportFn, ProjectAnnotations } from '@storybook/types';
 
 import { PreviewWeb } from './PreviewWeb';
 import {
@@ -56,7 +55,7 @@ const { history, document } = global;
 
 const mockStoryIndex = jest.fn(() => storyIndex);
 
-let mockFetchResult;
+let mockFetchResult: any;
 jest.mock('global', () => ({
   ...(mockJest.requireActual('global') as any),
   history: { replaceState: mockJest.fn() },
@@ -101,11 +100,12 @@ async function waitForSetCurrentStory() {
 }
 
 async function createAndRenderPreview({
+  // @ts-expect-error (not strict)
   importFn: inputImportFn = importFn,
   getProjectAnnotations: inputGetProjectAnnotations = getProjectAnnotations,
 }: {
-  importFn?: ModuleImportFn;
-  getProjectAnnotations?: () => WebProjectAnnotations<Framework>;
+  importFn?: Store_ModuleImportFn;
+  getProjectAnnotations?: () => ProjectAnnotations<Renderer>;
 } = {}) {
   const preview = new PreviewWeb();
   await preview.initialize({
@@ -146,6 +146,7 @@ describe('PreviewWeb', () => {
       const preview = new PreviewWeb();
       await expect(
         preview.initialize({
+          // @ts-expect-error (not strict)
           importFn,
           getProjectAnnotations: () => {
             throw err;
@@ -162,6 +163,7 @@ describe('PreviewWeb', () => {
       mockFetchResult = { status: 500, text: async () => err.toString() };
 
       const preview = new PreviewWeb();
+      // @ts-expect-error (not strict)
       await expect(preview.initialize({ importFn, getProjectAnnotations })).rejects.toThrow(
         'sort error'
       );
@@ -232,6 +234,7 @@ describe('PreviewWeb', () => {
     it('allows async getProjectAnnotations', async () => {
       const preview = new PreviewWeb();
       await preview.initialize({
+        // @ts-expect-error (not strict)
         importFn,
         getProjectAnnotations: async () => {
           return getProjectAnnotations();
@@ -514,6 +517,7 @@ describe('PreviewWeb', () => {
           renderToCanvas: undefined,
         });
         const preview = new PreviewWeb();
+        // @ts-expect-error (not strict)
         await expect(preview.initialize({ importFn, getProjectAnnotations })).rejects.toThrow();
 
         expect(preview.view.showErrorDisplay).toHaveBeenCalled();
@@ -627,6 +631,7 @@ describe('PreviewWeb', () => {
         });
 
         const preview = new PreviewWeb();
+        // @ts-expect-error (not strict)
         await preview.initialize({ importFn, getProjectAnnotations });
 
         await waitForRender();
@@ -682,7 +687,7 @@ describe('PreviewWeb', () => {
 
         const context = docsRenderer.render.mock.calls[0][0];
 
-        expect(context.componentStories().map((s) => s.id)).toEqual([
+        expect(context.componentStories().map((s: any) => s.id)).toEqual([
           'component-one--a',
           'component-one--b',
           'component-one--e',
@@ -892,6 +897,7 @@ describe('PreviewWeb', () => {
         document.location.search = '?id=component-one--a';
         componentOneExports.default.loaders[0].mockImplementationOnce(async () => gate);
 
+        // @ts-expect-error (not strict)
         await new PreviewWeb().initialize({ importFn, getProjectAnnotations });
         await waitForRenderPhase('loading');
 
@@ -953,6 +959,7 @@ describe('PreviewWeb', () => {
         document.location.search = '?id=component-one--a';
         projectAnnotations.renderToCanvas.mockImplementation(async () => gate);
 
+        // @ts-expect-error (not strict)
         await new PreviewWeb().initialize({ importFn, getProjectAnnotations });
         await waitForRenderPhase('rendering');
 
@@ -1032,6 +1039,7 @@ describe('PreviewWeb', () => {
         });
 
         document.location.search = '?id=component-one--a';
+        // @ts-expect-error (not strict)
         await new PreviewWeb().initialize({ importFn, getProjectAnnotations });
         await waitForRenderPhase('playing');
 
@@ -1099,6 +1107,7 @@ describe('PreviewWeb', () => {
 
           mockChannel.emit.mockClear();
           const story = await preview.storyStore.loadStory({ storyId: 'component-one--a' });
+          // @ts-expect-error (not strict)
           preview.renderStoryToElement(story, 'story-element' as any);
           await waitForRender();
 
@@ -1416,6 +1425,7 @@ describe('PreviewWeb', () => {
 
       document.location.search = '?id=component-one--a';
       projectAnnotations.renderToCanvas.mockImplementation(async () => gate);
+      // @ts-expect-error (not strict)
       await new PreviewWeb().initialize({ importFn, getProjectAnnotations });
       await waitForRenderPhase('rendering');
 
@@ -1512,6 +1522,7 @@ describe('PreviewWeb', () => {
       it('still renders the selected story, once ready', async () => {
         document.location.search = '';
         // We intentionally are *not* awaiting here
+        // @ts-expect-error (not strict)
         new PreviewWeb().initialize({ importFn, getProjectAnnotations });
 
         emitter.emit(SET_CURRENT_STORY, {
@@ -1618,6 +1629,7 @@ describe('PreviewWeb', () => {
           const preview = new PreviewWeb();
           // We can't wait for the initialize function, as it waits for `renderSelection()`
           // which prepares, but it does emit `CURRENT_STORY_WAS_SET` right before that
+          // @ts-expect-error (not strict)
           preview.initialize({ importFn, getProjectAnnotations });
           await waitForEvents([CURRENT_STORY_WAS_SET]);
 
@@ -1663,6 +1675,7 @@ describe('PreviewWeb', () => {
           const preview = new PreviewWeb();
           // We can't wait for the initialize function, as it waits for `renderSelection()`
           // which prepares, but it does emit `CURRENT_STORY_WAS_SET` right before that
+          // @ts-expect-error (not strict)
           preview.initialize({ importFn, getProjectAnnotations });
           await waitForEvents([CURRENT_STORY_WAS_SET]);
 
@@ -1707,6 +1720,7 @@ describe('PreviewWeb', () => {
           const preview = new PreviewWeb();
           // We can't wait for the initialize function, as it waits for `renderSelection()`
           // which prepares, but it does emit `CURRENT_STORY_WAS_SET` right before that
+          // @ts-expect-error (not strict)
           preview.initialize({ importFn, getProjectAnnotations });
           await waitForEvents([CURRENT_STORY_WAS_SET]);
 
@@ -2027,6 +2041,7 @@ describe('PreviewWeb', () => {
           componentOneExports.default.loaders[0].mockImplementationOnce(async () => gate);
 
           document.location.search = '?id=component-one--a';
+          // @ts-expect-error (not strict)
           await new PreviewWeb().initialize({ importFn, getProjectAnnotations });
           await waitForRenderPhase('loading');
 
@@ -2060,6 +2075,7 @@ describe('PreviewWeb', () => {
 
           document.location.search = '?id=component-one--a';
           projectAnnotations.renderToCanvas.mockImplementation(async () => gate);
+          // @ts-expect-error (not strict)
           await new PreviewWeb().initialize({ importFn, getProjectAnnotations });
           await waitForRenderPhase('rendering');
 
@@ -2094,6 +2110,7 @@ describe('PreviewWeb', () => {
           componentOneExports.a.play.mockImplementationOnce(async () => gate);
 
           document.location.search = '?id=component-one--a';
+          // @ts-expect-error (not strict)
           await new PreviewWeb().initialize({ importFn, getProjectAnnotations });
           await waitForRenderPhase('playing');
 
@@ -2147,6 +2164,7 @@ describe('PreviewWeb', () => {
           componentOneExports.a.play.mockImplementationOnce(async () => gate);
 
           document.location.search = '?id=component-one--a';
+          // @ts-expect-error (not strict)
           await new PreviewWeb().initialize({ importFn, getProjectAnnotations });
           await waitForRenderPhase('playing');
 
@@ -2550,6 +2568,7 @@ describe('PreviewWeb', () => {
         mockFetchResult = { status: 500, text: async () => err.toString() };
 
         const preview = new PreviewWeb();
+        // @ts-expect-error (not strict)
         await expect(preview.initialize({ importFn, getProjectAnnotations })).rejects.toThrow(
           'sort error'
         );
@@ -2570,6 +2589,7 @@ describe('PreviewWeb', () => {
         mockFetchResult = { status: 500, text: async () => err.toString() };
 
         const preview = new PreviewWeb();
+        // @ts-expect-error (not strict)
         await expect(preview.initialize({ importFn, getProjectAnnotations })).rejects.toThrow(
           'sort error'
         );
@@ -3073,6 +3093,7 @@ describe('PreviewWeb', () => {
         await waitForEvents([STORY_MISSING]);
 
         mockChannel.emit.mockClear();
+        // @ts-expect-error (not strict)
         preview.onStoriesChanged({ importFn, storyIndex });
         await waitForRender();
         expect(mockChannel.emit).toHaveBeenCalledWith(STORY_RENDERED, 'component-one--a');
@@ -3083,6 +3104,7 @@ describe('PreviewWeb', () => {
       const newStandaloneDocsExports = { default: jest.fn() };
 
       const newImportFn = jest.fn(async (path) => {
+        // @ts-expect-error (not strict)
         return path === './src/Introduction.mdx' ? newStandaloneDocsExports : importFn(path);
       });
 
@@ -3092,6 +3114,7 @@ describe('PreviewWeb', () => {
         mockChannel.emit.mockClear();
         docsRenderer.render.mockClear();
 
+        // @ts-expect-error (not strict)
         preview.onStoriesChanged({ importFn: newImportFn });
         await waitForRender();
 
@@ -3111,6 +3134,7 @@ describe('PreviewWeb', () => {
         const preview = await createAndRenderPreview();
         mockChannel.emit.mockClear();
 
+        // @ts-expect-error (not strict)
         preview.onStoriesChanged({ importFn: newImportFn });
         await waitForRender();
 
@@ -3128,6 +3152,7 @@ describe('PreviewWeb', () => {
         const preview = new PreviewWeb();
         await expect(
           preview.initialize({
+            // @ts-expect-error (not strict)
             importFn,
             getProjectAnnotations: () => {
               throw err;
@@ -3148,6 +3173,7 @@ describe('PreviewWeb', () => {
         const preview = new PreviewWeb();
         await expect(
           preview.initialize({
+            // @ts-expect-error (not strict)
             importFn,
             getProjectAnnotations: () => {
               throw err;
@@ -3190,6 +3216,7 @@ describe('PreviewWeb', () => {
       expect(mockChannel.emit).toHaveBeenCalledWith(CONFIG_ERROR, err);
     });
 
+    // @ts-expect-error (not strict)
     const newGlobalDecorator = jest.fn((s) => s());
     const newGetProjectAnnotations = () => {
       return {
@@ -3319,6 +3346,7 @@ describe('PreviewWeb', () => {
       const [gate, openGate] = createGate();
       componentOneExports.a.play.mockImplementationOnce(async () => gate);
       const preview = new PreviewWeb();
+      // @ts-expect-error (not strict)
       await preview.initialize({ importFn, getProjectAnnotations });
       await waitForRenderPhase('playing');
 
@@ -3341,6 +3369,7 @@ describe('PreviewWeb', () => {
       mockChannel.emit.mockClear();
       const [gate, openGate] = createGate();
       componentOneExports.b.play.mockImplementationOnce(async () => gate);
+      // @ts-expect-error (not strict)
       preview.renderStoryToElement(
         await preview.storyStore.loadStory({ storyId: 'component-one--b' }),
         {} as any
@@ -3373,6 +3402,7 @@ describe('PreviewWeb', () => {
       const preview = new PreviewWeb();
       await expect(
         preview.initialize({
+          // @ts-expect-error (not strict)
           importFn,
           getProjectAnnotations: () => {
             throw err;
@@ -3388,6 +3418,7 @@ describe('PreviewWeb', () => {
       mockFetchResult = { status: 500, text: async () => err.toString() };
 
       const preview = new PreviewWeb();
+      // @ts-expect-error (not strict)
       await expect(preview.initialize({ importFn, getProjectAnnotations })).rejects.toThrow(
         'sort error'
       );
@@ -3398,11 +3429,13 @@ describe('PreviewWeb', () => {
     it('waits for stories to be cached', async () => {
       const [gate, openGate] = createGate();
 
+      // @ts-expect-error (not strict)
       const gatedImportFn = async (path) => {
         await gate;
         return importFn(path);
       };
 
+      // @ts-expect-error (not strict)
       const preview = await createAndRenderPreview({ importFn: gatedImportFn });
 
       let extracted = false;