+++ code/renderers/react/src/public-api.tsx
@@ -8,16 +8,16 @@ import type {
 import { start } from '@storybook/core-client';
 
 import { renderToCanvas, render } from './render';
-import type { ReactFramework } from './types';
+import type { ReactRenderer } from './types';
 
-interface ClientApi extends Addon_ClientStoryApi<ReactFramework['storyResult']> {
+interface ClientApi extends Addon_ClientStoryApi<ReactRenderer['storyResult']> {
   configure(loader: Addon_Loadable, module: NodeModule): void;
   forceReRender(): void;
   raw: () => any; // todo add type
 }
 const FRAMEWORK = 'react';
 
-const api = start(renderToCanvas, { render });
+const api = start<ReactRenderer>(renderToCanvas, { render });
 
 export const storiesOf: ClientApi['storiesOf'] = (kind, m) => {
   return (api.clientApi.storiesOf(kind, m) as ReturnType<ClientApi['storiesOf']>).addParameters({
@@ -29,4 +29,4 @@ export const configure: ClientApi['configure'] = (...args) => api.configure(FRAM
 export const forceReRender: ClientApi['forceReRender'] = api.forceReRender;
 export const raw: ClientApi['raw'] = api.clientApi.raw;
 
-export type DecoratorFn<TArgs = Args> = DecoratorFunction<ReactFramework, TArgs>;
+export type DecoratorFn<TArgs = Args> = DecoratorFunction<ReactRenderer, TArgs>;