+++ code/renderers/svelte/src/render.ts
@@ -1,14 +1,15 @@
 /* eslint-disable no-param-reassign */
 import type { Store_RenderContext, ArgsStoryFn } from '@storybook/types';
 import type { SvelteComponentTyped } from 'svelte';
+
 // eslint-disable-next-line import/no-extraneous-dependencies
 import PreviewRender from '@storybook/svelte/templates/PreviewRender.svelte';
 
-import type { SvelteFramework } from './types';
+import type { SvelteRenderer } from './types';
 
-const componentsByDomElement = new Map<SvelteFramework['canvasElement'], SvelteComponentTyped>();
+const componentsByDomElement = new Map<SvelteRenderer['canvasElement'], SvelteComponentTyped>();
 
-function teardown(canvasElement: SvelteFramework['canvasElement']) {
+function teardown(canvasElement: SvelteRenderer['canvasElement']) {
   if (!componentsByDomElement.has(canvasElement)) {
     return;
   }
@@ -28,8 +29,8 @@ export function renderToCanvas(
     showError,
     storyContext,
     forceRemount,
-  }: Store_RenderContext<SvelteFramework>,
-  canvasElement: SvelteFramework['canvasElement']
+  }: Store_RenderContext<SvelteRenderer>,
+  canvasElement: SvelteRenderer['canvasElement']
 ) {
   const existingComponent = componentsByDomElement.get(canvasElement);
 
@@ -67,7 +68,7 @@ export function renderToCanvas(
   };
 }
 
-export const render: ArgsStoryFn<SvelteFramework> = (args, context) => {
+export const render: ArgsStoryFn<SvelteRenderer> = (args, context) => {
   const { id, component: Component } = context;
   if (!Component) {
     throw new Error(